requirements.txt
instagrapi>=2.0.0
pandas>=2.1.3
openpyxl>=3.1.2
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.7.0
loguru>=0.7.2
pydantic>=1.10.9
python-jose>=3.3.0
cryptography>=41.0.7
schedule>=1.2.1
tqdm>=4.66.1 